pyproject.toml
[project]
name = "deft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "torch==2.4.0+cu118; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "torch==2.4.0; sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'aarch64')",
    "torchvision==0.19.0+cu118; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "torchvision==0.19.0; sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'aarch64')",
    "numpy==1.23.5",
    "wandb>=0.18.3",
    "omegaconf>=2.3.0",
    "matplotlib>=3.9.2",
    "einops>=0.8.0",
    "tqdm>=4.66.5",
    "hydra-core>=1.3.2",
    "gdown>=5.2.0",
    "lmdb>=1.5.1",
    "ema-pytorch>=0.7.3",
    "odl>=0.7.0",
    "dival>=0.6.1",
    "clean-fid>=0.1.35",
    "torchmetrics>=1.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
find-links = [
    "https://download.pytorch.org/whl/cu118/torch",
    "https://download.pytorch.org/whl/cu118/torchvision",
]
dev-dependencies = [
    "ipykernel>=6.29.5",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
